requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.3
decorator==5.1.1
distro==1.9.0
duckdb==0.9.2
executing==2.0.1
faiss-cpu==1.7.4
Faker==19.13.0
fonttools==4.48.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
install==1.3.5
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langchain-experimental==0.0.50
langchain-mistralai==0.0.3
langsmith==0.0.87
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistralai==0.0.11
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.7.0
orjson==3.9.12
packaging==23.2
pandas==1.5.3
pandasai==1.5.19
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pydantic==1.10.14
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pypdf==3.17.4
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.16.2
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
streamlit==1.30.0
streamlit-pills==0.3.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
